[HandshakeFixArgName] Match arg names from source
This introduces an experimental pass that operates on Handshake-level IR
but also takes as input the source C file from which the IR was
originally created. The pass parses the C file containing the kernel
declaration and attemps to extract the kernel's argument names from it.
It then replaces the argument names of the corresponding
handshake::FuncOp operation in the IR. This is useful for using the
legacy HLS verifier downstream, which expects argument names in VHDL to
match those in the source code.

This commit also includes a small modification to the DOT printer to
make it use the same names for memories as the original source code.