[dynamatic] New Dynamatic frontend
This commit introduces the first version of the frontend for Dynamatic.
It somewhat mimics the command syntax and logic of legacy Dynamatic's
frontend but can also automatically simulate/verify and logic-synthesize
code after compilation/synthesis. At this point, the frontend still
relies on a couple legacy tools (dot2vhfl/HLS verifier) for which it is
required that legacy Dynamatic is installed somewhere on the system.
This will eventually disappear once we have our own version of this part
of the toolchain.

Some sample sripts for the frontend are provided to help users get an
idea of how to use the currently available commands.