[Buffers] Split common logic from specific MILP definitions
This is a significative refactoring of the buffer placement MILP to
improve abstractions and code-reuse opportunities. The
`BufferPlacementMILP` class is now abstract and only serves to factor in
the parts of the buffer placement logic and state management that smart
buffer placement implementations are likely to need. Our first placement
implementation previously contained fully in this class now lives, for
all its algorithm-specific parts in the
`dynamatic::buffers::fpga20::FPGA20Buffers` class. The latter defines
the specific MILP variables, constraints, and objective that
characterize this specific MILP, as well as provides the logic to
extract buffer placement decisions from the MILP's results.