[Buffer] Add pass parameter to choose algorithm
This adds a parameter to the `HandshakePlaceBuffers` pass that allows
the user to control which algorithm to use for buffer placement. Current
choices are "fpga20" and "fpga20-legacy", which optimize the same MILP
but translate its solution into different placement decisions (controls
the `legacyPlacement` flag of `FPGA20Buffers`). This will be useful in
the future when adding new placement algorithms.