[CFDFC] CFDFC unions and disjoint-set computation
This commit adds a new data type representing a union of CFDFCS
(`CFDFCUnion`) as well as a method to compute, from a list of CFDFCs,
their disjoin-set unions (based on the blocks they cover) using a
rank-based union algorithm. This is immediately useful for FPL'22-style
buffer placement, which will be implemented soon.