[Buffers] Placement works with single-block functions
This makes the `HandshakePlaceBuffersPass` work correctly on functions
with a single block (and as a consequence no basic block arch).
Previously, the first CFDFC extraction MILP would fail on those
situations because of an empty objective. To fix this we simply do not
attempt to extract CFDFCs if the function has no archs, and let the
buffer placement MILP execute trivially (all terms of the objective to
maximize will be negative so it results in 0 buffers being placed).

We take this opportunity to slightly improve error reporting around
Gurobi models (both for CFDFC extraction and buffer placement) to make
error messages more readable and informative.