[Buffers] Support for channel-specific combinational delays (#43)
This commit (that includes a CIRCT change) adds support for
channel-specific combinational delays during buffer placement (see
related issue for an explanation of the implemented delay model).

This change includes both a modification of the `ChannelBufPropsAttr` IR
attribute and of its related infrastructure to support attaching these
delays to IR objects and modifications to the FPGA'20 MILP to take those
into account during placement (in the path constraints).