[NameAnalysis] New analysis managing unique op names
This commit starts to address #45 by introducing an analysis
data-structure (`NameAnalysis`) meant to be loaded by Dynamatic passes
to query and set unique names for all operations in the IR, as well as
verify the uniqueness and persistence of names throughout the IR's
im-memory lifetime. A next commit will make most current Dynamatic
passes use this analysis internally to ensure that any operation they
introduce is named.