[NameAnalysis] Add standalone name getters and fix bugs
This commit adds two simple functions to query an operation's and
operand's unique name, so that one does not have to instantiate the full
`NameAnalysis` (which walks the entire IR) to simply attempt to get the
name of an operation/operand. Their implementation is very similar to
the analysis's getters, but directly go read the operation attribute
instead of doing a map lookup.

This also fixes two simple issues in the `NameAnalysis` implementation.