Make most passes subclasses of `DynamaticPass` (#45)
This commit makes trivial and systematic changes to most Dynamatic
passes to make them inherit from the new `dynamatic::DynamaticPass`
abstract class instead of directly from
`mlir::OperationPass<mlir::ModuleOp>` (which the former inherits from).
This is prompted by the introduction of the naming analysis
infrastructure that assigns unique names to all IR operations (and
verifies the uniqueness of names at all times), which the vast majority
of Dynamatic passes must leverage. To avoid boilerplate code duplication
the `dynamatic::DynamaticPass` class implements this verification logic
and acts as a wrapper around the pass-specific logic. In the future,
invariants commonly encountered in Dynamatic passes (e.g., IR
materialization) will be added to this wrapper to make pass
pre/post-condition checking more consistent across passes and easier for
users to add to their passes.

In order to avoid breaking all unit tests, this commit also introduces a
new pass removing all unique names (`RemoveOperationNames`) from the IR,
which is ran right after the pass-under-test in all FileCheck-parsable
files. We do not care to have visible operation names in unit tests,
which just clutter the output and are not an intrinsic part of the
behavior one wishes to test for. A reciprocal `NameAllOperations` pass
is introduced, whose behavior is self-explanatory.

This finally closes #45.