[OperandAttribute] Support for generic operand attributes (#46)
This commit adds support for working with operanf attributes: attributes
that are semantically attached to MLIR operands rather than operations.
Since this is not natively supported by MLIR, we introduce support
functions to allow one to work with operand attributes with almost the
same ease as regular attributes.

Dynamatic already used "channel attributes" prior to this change in the
form of channel-specific buffering properties. This commit, however,
1. makes the implementation general to work with any user-provided
   operand attribute type.
2. makes operation attributes attach of the channel's "owning" operation
   (the one which the channel is an operand of) instead of the
   "defining" operation.
3. moves the templated implementation to a `Support` header.
The original/partial implementation of operand attributes is thus
deleted in favor of the new one.