[BackAnnotate] Generic IR back-annotation pass (#47)
This commit adds a back-annotation pass to Dynamatic, which allows one
to set attributes on IR elements (operations or operands) from a
JSON-formatted back-annotation file with a specific syntax (see
documentation in pass header). This is very useful when one gains
insights about the IR from some external program (e.g., from Vivado
after logic synthesis) and would like to go back to the IR and set
annotations that will ultimately be leveraged by tools/passes to yield a
better final dataflow circuit.

At this point, only a single "proof-of-concept" annotation type is
supported by the pass; it corresponds to the ChannelBufPropsAttr
operand attribute used during buffer placement. It should be relatively
easy to add new annotation types in the future, though the method for
doing so could use some improvements (see documentation in pass
implementation).

Closes #47.