[visual-dataflow] DOT/CSV Parsing (#40)
## Description
This pull request introduces the initial version of parsing tools that
will help us represent a specific circuit graphically in Godot and track
its state evolution for each cycle.

The key components of this PR include:
- *Parser*: This class serves as the main parsing interface, allowing
developers to process both .dot and .csv files.
- *DotParser*: The primary purpose of this component is to parse .dot
files in order to create a Graph object representing a specific circuit.
It also stylizes the resulting graph in a
[Graphviz](https://graphviz.org/) style (position of nodes and edges).
- *CSVParser*: Designed to handle .csv files, this parser identifies the
state of each edge for each cycle.
- *ReformatDot*: A helper class that streamlines the formatting of .dot
files, making them easier to parse and ensuring consistency.
- *GraphComponents*: This module defines the data structures used to
create a Graph object.

This PR successfully parses and processes data samples provided in the
current implementation. However, it's important to note that further
adaptation may be necessary when integrating this functionality with
GDExtension for the graphical user interface. ReformatDot can be further
optimized to reduce its line count.

## What’s new 
- Parser, DotParser, CSVParser, ReformatDot, and GraphComponents
classes.
- Defined the main parsing and graph construction logic.
- Ensured proper styling of the graph in Graphviz style (theoretically).

## Testing
- Verified the functionality and accuracy of parsed data by testing with
sample .dot and .csv files.

---------

Co-authored-by: Alice Potter <alice.potter@epfl.ch>
Co-authored-by: alicepotter <alicepotter02@gmail.com>
Co-authored-by: Quentin Gross <quentin.gross@epfl.ch>
Co-authored-by: Lucas Ramirez <lucas.rami@proton.me>