[Handshake] Refactor `MemoryControllerOp`'s definition (#41)
This commit makes working with `handshake::MemoryControllerOp`
operations significantly easier by simplifying its definition and
introducing data structures to structurally represent the logical ports
connected to the memory interface (in CIRCT). This leads to shorter,
more robust, and more readable code wherever the now deleted `accesses`
intrinsic attribute previously needed to be parsed.

To make this refactoring possible, memory controller operations now
require that the "port operations" producing the SSA values that end up
as its inputs are explicitly annotated with a `bb` attribute indicating
the basic block they used to belong to. This required the following
changes in the codebase, in addition to the refactoring mentionned
above:
- Deletion of the `id-basic-blocks` parameter in the
  `StandardToHandshakeFPGA18` conversion pass since basic block tagging
  is now mandatory. The pass now operates as if this flag was always
  set.
- Modification of some unit tests to account for the removal of
  `MemoryControllerOp`'s intrinsic attributes as well as the conversion
  pass change discussed above.
- Modification to the synthesis frontend script to remove the
  `id-basic-blocks` parameter to the elastic pass.

This mostly addresses #41 minus the text operation format change, which
will be treated separately at a later time.