[LSQ] Introduce `handshake::LSQOp` operation
This commit adds support for the new load-store-queue operation
(`circt::handshake::LSQOp`) in Dynamatic. Since it is part of the
Handshake dialect the main changes of this commit are in CIRCT. Some of
the APIs related to memory interfaces in general changed along the
introduction of the new interface therefore there are some syntactic
changes in the parts of the codebase that are parsing memory ports.