[LSQ] Annotations to drive memory interface placement (#50)
This commit continues adding support for LSQs in Dynamatic by
introducing a new `NOLSQAttr` attribute meant to be attached to memory
operations (from the affine/memref dialects) in the IR that do not need
to go through an LSQ. This attribute is meant to be read by the
cf-to-handshake conversion pass to drive the placement and connectivity
of memory interfaces for each memory region.

Eventually, this attribute will be set by our memory analysis pass which
will determine which memory accesses are (in)dependent from others. For
now however, we introduce the `ForceMemoryInterface` which enables
corse-grain control of this attribute over the entire IR, and thus can
force the exclusive use of one type of memory interfaces (MCs or LSQs)
in our dataflow circuits once the cf-to-handshake conversion pass
includes the logic to place LSQs.

Closes #50.