[Support] Always load the Handshake dialect with `DynamaticPass`es
This makes the helper TableGen pass declaration class `DynamaticPass`
automatically register the `HandshakeDialect` as a dependent dialect. We
always need it because our naming analysis, loaded automatically by
every `DynamaticPass`, uses the `handshake::NameAttr` attribute
internally which is part of the Handshake dialect.

Previously, if nothing was part of the Handshake dialect in the input IR
of a pass that did not declare the Handshake dialect dependent, then the
pass would eventually crash when the naming analysis would attempt to
name all unnamed operations before exiting the pass, because the dialect
would not have been loaded in the context.

This change means that additional dependent dialects need to be passed
to `DyanmaticPass` instead of declared with `let` in child passes. This
is slightly less elegant but AFAIK is the only wat to achieve this
behavior.