[NameMemoryOps] Delete pass (superseeded by `NamingAnalyis`)
This deleted the `NameMemoryOps` pass rendered useless by our recent
addition of the `NamingAnalyis`, which constantly maintains unique names
for all operations instead of just memory operations. Parts of the
codebase that used the pass along with the attribute it manipulated are
refactored to now use the `NamingAnalysis`.

Both the `--name-memory-ops` and `--analyze-memory-accesses` passes are
deleted from our frontend compilation script. The former because it no
longer exists, the second because it at the moment stores useless
information in the IR that just cluters it (and that may never be
actually needed, even when we start using the pass for LSQ placement).