[Support] Helpers to copy attributes between ops
This introduces a pair of helper functions in the former
`OperandAttribute.h` header, now renamed to `Attribute.h`, to copy
an attribute from a certain type (if present) from a source operation to
a destination operation. Passes that used to do this manually are made
to use this new helper function.