[StandardToHandshakeFPGA18] Simplify memmory managenment data structures
This is a small change that just swaps vectors for map-vectors in the
internal data structures used to track memory interfaces and their
corresponding memory accesses. This preserves the same deterministic
ordering when iterating over map values (which is why I originally used
vectors before I realized that ordered maps were a thing). This removes
quite a bit of boilerplate code.