[Buffers] Add simple placement strategy to placement pass
This commit adds a buffer placement strategy roughly equivalent to the
one that we used to use from CIRCT in our own buffer placement pass. The
strategy simply places one opaque slot and one transparent slot on each
output channel of each merge-like operation in the IR, ensuring
correctness and a certainly terrible performance.

This change involves moving some declarations around and renaming a
header file, hence the large number of files changed. Of siginificance,
the commit improves the algorithm-selection infrastructure in the buffer
placement pass so that algorithms that require Gurobi are clearly
separated from those that do not (which includes the one we just
introduced) in a way that work nicely with the Gurobi-conditional
compilation.

This change was required to make this placement strategy work with the
refactor of memory interface that now require every port operation to
explicitly belong to a block, which the previous implementation did not
care about. This also centralizes all of our buffer placement needs,
which is a nice touch.