[IntegrationTest] Automate profiling input generation
This commit automates the generation of profiling inputs at the cf level
for the buffer placement pass, which is infinitely more convenient than
having to manually create a file containing inputs. As such, all
`inputs.txt` files (one per integration test) are removed.

Our new approach relies on a C macro (CALL_KERNEL) defined in a utility
file that all integration tests should include. The macro wraps a call
to the kernel under evaluation and, depending on whether a compile-time
macro is defined, will either expand to a simple function call or will
first dump all of the kernel's argument to standard output before
actually calling the kernel. The latter works using some heavily
templated C++ code that Polygesit or the HLS verifier cannot compile,
therefore we ensure that, when compiling the code with no intent to
profile it, the latter remains in the realm of pure C.

This change is integrated into the synthesis frontend script (dumping
the kernel inputs to a `profiler-inputs.txt` file), which is also reworked
(along with the simulate script and the frontend itself) to internally
work with absolute paths instead of relative ones, which is more robust.

The commit also remvoes the `src` subfolder within `integration-test`
which had no real purpose. All integration tests are now stored directly
under `integration-test`.