[dynamatic-opt] Make our tool register standard passes we care about
This commit makes our dynamatic-opt tool also register all the standard
passes we are likely to care about for the foreseeable future. This
makes it so that all the steps in our optimization/lowering flow rely
solely on our optimizer instead of a mix of the standard one (mlir-opt)
and ours, which is gonna be more convenient as the frontend becomes more
fully-fledged and offers more fine-grained control over the passes one
can run from it. The current synthesis script is updated to rely only on
dynamatic-opt.

As a logical consequence, this commit removes the symlinks of mlir-opt
and circt-opt that the build script used to create in the `bin`
directory.