[dynamatic] Slight change in compilation script flow
This touches on two passes in the flow specified by our frontend script
that performs the lowering from C to DOT.

First, the --flatten-memref-row-major pass is moved earlier in the flow
since there is no disadvantage in doing so but, in the future, it may
allow optimization of memory address computations. Second, the
--arith-strength-reduction pass is parameterized explicitly since it is
not clear that replacing multiplications with large adder/shift trees is
so advantageous.