[NameAnalysis] More convenient get/set API
This commit improves the get/set API of our name analysis to make it
less "failable" from the client code perspective. This means that less
methods can return failures, instead adopting some default meaningful
behavior that is considered a "success". This avoids cumbersome
error-checking in user code and leads to shorter/more readable code
overall.

This is achieved mainly by changing the behavior of `getName` so that it
now allocates a new unique name on the operation if it doesn't have one
when the call is made. This makes the getter unfailable, which
reverberates across other methods, making them also unfailable.

At the same time, this commit slightly improves the analysis's
reliability with intrinsically named operations. It is now less likely
that one "namable" operation will randomly get the same name as one of
those intrinsically named ones.