[MarkMemoryDependencies] Cleanup memory analysis pass
This commit cleans up the memory analysis pass in preparation for
connecting its results to LSQ placement. Behavior-wise the pass should
remain identical except in the treatment of non-affine memory access
pairs for which the pass's behavior is fixed when the source and
destination accesses are the same operation. The pass is also renamed
from `AnalyzeMemoryAccesses` to `MarkMemoryDependencies`.