[MarkMemoryInterfaces] Mark required interfaces after mem. analysis
This commit introduces a new pass (`--mark-memory-interfaces`) that uses
the results of the polehdral memory analysis
(`--mark-memory-dependencies`) to mark memory interfaces with an
attribute indicating the kind of memory interfaces they should
eventually connect to. These are later read by the cf-to-handshake
conversion pass to drive the placement of memory interfaces.

This essentially completes a basic and usable LSQ flow (only missing
more complex flow analysis to reduce LSQ usage), from memory analysis
to smart placement.