[Support] Remove `DynamaticPass`'s template parameter
This is a very simple change which reverberates in many files since it
removes the template parameter of the base pass class that the vast
majority of Dynamatic passes derive from.

The template parameter (which controlled a pre-pass check that all IR
operations were uniquely named) achieved nothing that cannot be achieved
in other simpler ways and was very verbose. Eventually, its behavior
(and similar ones) could be controlled via a pass constructor or a
virtual method that implementors could override. At the moment though,
it is simply removed in favor of a default behavior that doesn't perform
the verification it controlled, since no pass but the buffer placement
one cares about it. The latter now makes sure to name IR operations
itself before executing the bulk of its logic.