[Handshake] Separate access ports for each mem. interface
This commit simply reflects modification in the definitions of Handhsake
operations in our CIRCT fork. The `DynamaticLoadOp` and
`DynamaticStoreOp` operations have been replaced by
memory-interface-specific versions specialized for MCs (`MCLoadOp`,
`MCStoreOp`) and LSQs (`LSQLoadOp`, `LSQStoreOp`).

This is motivated by the fact that the timing models and RTL
implementations of access ports is interface-dependent. Having a single
pair of load and store operations failed to capture these differences.
Thanks to the introduction of the `LoadOpInterface` and
`StoreOpInterface` which expose the common interfaces between both load
ports and both store ports, respectively, only minor changes to the
codebase were necessary.