[ArithReduceStrength] Improved optimization of x * (-1) pattern
This commit improves the strength reduction rewrite pattern that
replaces the x * (-1) + y pattern with y - x, removing a multiplication
in the process.

This pattern used to only be able to match on additions where one
operand was of the form x * (-1), however, it works almost equally well
for substractions, which we now support. For substractions, if the
x * (-1) pattern is on the LHS, then a sign flip is required after
replacement with an addition. The resulting structure is still cheaper
than a multiplier in terms of area. New unit tests are added to check
for the extra optimization scenarios we now support.