[Support] Generic MILP API and state management
This commit refactors our MILP-related code to separate the
buffer-placement-specific logic from the generic MILP management
infrastructure, which can be used in different contexts (e.g., for CFDFC
extraction). A new `MILP` abstract support class is added to centralize
the latter, which the existing `BufferPlacementMILP` class now inherits.

This makes the code cleaner, better encapsulated, and generally enables
a better separation of concerns between the various classes. Currently
the `MILP` class and its descendants are all conditionally compiled
based on the presence of Gurobi, but in the future one can imagine that
the `MILP` class will offer a solver-agnostic API to work with MILPs.