[Buffers] Remove unused `maxPeriod` class member
Simply removes references to the currently unused `maxPeriod` argument
in the `FPGA20Buffers` class. This was previously used as a constant
upper-bound in some MILP constraints to "cancel out" some terms, however
it had already been replaced with a value computed on the fly instead of
one set during object construction.