[export-dot] Fix width of constants' "value" annotation (#54)
This is a simple change that fixes the text width of the "value"
annotation's value for constant operations so that it correlates
correctly with the bitwidth of the node's output in the DOT. This
matters in legacy modes because the dot2vhdl tool parses this annotation
to determine what to print in the generated VHDL for the constant value.

Fixes #54.