[data] Replace 100.0 value by 0 in unit timing nodels
This commit replaces the "special value" 100.0 given to most VR, CV, CR,
VC, and VD fields by 0. In legacy Dynamatic, this special value was used
to denote an "open path" in mixed domain constraints for FPL'22 buffer
placement. However, since mixed domain paths are anyway hardcoded in the
legacy implementation--and will be in the new version of the pass in a
first time, until we have proper timing models--the timing models do
not need to also model this. It in facts complicate the implementation
since one has to check for this special value, whereas 0 can be
considered an acceptable default value when one of these paths' delay is
queried directly.