[Buffers] Accurate buffering constraints on LSQ control path
This commit significantly improves the logic for identifying channels on
the control path to LSQs that must contain at least one opaque buffer
slot to prevent simultaneous group allocations, which cause deadlocks.
The effect of this change is two-fold:
1. It avoids inadvertently forcing an opaque buffer between forks and
   LSQ control ports and yielding unsatisfiable buffering constraints
   for such channels, making buffer placement fail during pre-condition
   checking.
2. It avoid forcing the placement of opaque buffers on fork outputs that
   are not part of the LSQ control path (for forks which trigger an LSQ
   group allocation). This ensures that a buffer is requested *only* on
   the control path to another group allocation to the same LSQ.

This change seems to eliminate the deadlocks that happened in a couple
of our benchmarks.