[Buffers] FPL'22 smart buffer placement
This commit introduces our best buffer placement algorithm yet,
originally described in the following papper published at FPL'22:
https://ieeexplore.ieee.org/abstract/document/10035122.

This new MILP-based approach to buffer placement works very similarly to
our existing one but is much more complete in terms of modelling.
Indeed, where our previous approach only considered delays and cycles on
the data timing domain, this new approach additionally considers them
on the valid and ready domain, as well as in-between domains (so called
mix-domains). This makes the buffer placement it derives much more
"informed" than our previous approach, and should improve the
performance of our circuits in execution time and critical path delay.
Another significant difference is that, instead of solving at MILP for
all CFDFCs of a function at the same time, this approach first creates
disjoit set unions of the function's CFDFCs and then solves an
independent MILP for each of them.

To reduce code size and due to the similarity to FPGA'20 buffer
placement, a lot of logic previously part of the `FPGA20Buffers` class
is generalized and moved inside the `BufferPlacementMILP` class, so that
both algorithms (both subclasses of the latter class) can use it.

Some works remains to fully complete the algorithm:
1. Another smaller MILP needs to be solved at the end to buffer
   potentially buffer channels that are not part of any CFDFC.
2. Currently the MILP's path constraints ignore channel-specific
   combinational delays (supported by FPGA'20 buffer placement). These
   should be refactored to take them into account.

This moves forward the effort described in #16.