[Buffers] Fix bug when `dump-logs` flag is not provided
This fixes a bug in the `HandshakePlaceBuffers` pass wherein the pass
would crash if it was not launched with the `dump-logs` flag
(instructing it to log placement information as well as a descripton
of all solved MILPs to disk as it progresses). In the abscence of such
flag the `logger` variable passed to the MILPs' constructor from the
pass would be a null reference, causing null pointer dereference issues
down-the-line.

The solution is to simply call the version of the MILPs constructor that
do not take a logger when the pass is ran without the `dump-logs` flag.