[Buffers] Add channel delay support to FPL'22 placement
This commit completes the FPL'22 buffer placement MILP by introducing
support for channel-specific delays (as specified in channel buffering
properties) in the expression of the MILP's path constraints. The change
is semantically identical to what was introduced for FPGA'20 buffer
placement in e113d27 (issue #43). As such the existing code for channel
path constraints is generalized and moved to the `BufferPlacementMILP`
class, making it usable by both placement strategies.

This moves forward on #16. After more extensive testing, the new buffer
placement algorithm will become the default one in our compile flow.