[Support] Add Handshake-IR helpers
This commit adds a new support file meant to contain heterogeneous
analysis/processing functions that operate on Handshake-level IR. This
is a sort of "dump" for free-standing functions not logically associated
to any pass and potentially useful in various contexts. Two functions
from the Handshake canonicalization pass are already moved there and
another one related to LSQ control paths is added in preparation for the
introduction of a new and improved materialization pass.