[HandshakeMaterialize] New materialization pass
This commit introduces the `HandshakeMaterialize` pass, which replaces
(and improves over) the `HandshakeMaterializeForksSinks` pass (from
CIRCT) in our compile flow. The pass ensures that every value in
Handshake functions is used exactly once, which is a prerequisite for
some passes and for RTL export in general.

The pass materializes an entire MLIR module by applying a sequence of
three transformation steps on each Handshake function in the module.
1. First, eager forks and sinks are inserted into the function as needed
   to ensure that every value is used exactly once. This should behave
   in exactly the same way as the CIRCT pass.
2. Second, unnecessary forks and sinks are optimized away. Unnecessary
   forks and sinks may appear after the first step if the input IR was
   already partially materialized. This is implemented as a set of
   operation rewrite patterns applied greedily, which partially existed
   in CIRCT but were rewritten for the new pass.
3. Finally, some eager forks on the memory control paths are transformed
   into lazy forks to ensure that no LSQ receives more than one group
   allocation during any given cycle, which is known to cause deadlocks.
   This is the proper way of working with LSQs and was never implemented
   fully before. Since this is the first time we use lazy forks in
   Dynamatic (despite their existence as a Handshake operation for a
   long time), this commit also adds support for them in the
   `DOTPrinter` so that they are parsable by legacy Dynamatic's
   toolchain.

This commit also drop references to the CIRCT helper function that
checks whether every value in a Handshake function has exactly one use
in favor of an equivalent function implemented directly in Dynamatic.

Comprehensive unit tests are included for the new pass.