[Buffers] Fix channel path constraints logic
This commits fixes the definition of one type of channel path
constraint involved in our MILP-based buffer placement algorithms.
Previously, the linear expression was missing a binary variable which
determines whether one of the terms of the linear expression is taken
into account based on the presence of a buffer on the channel.

This also includes some cosmetic changes to the definition of other
channel path constraints to make them more readable overall.