[Buffers] Do not "compensate" results in FPL'22 placement
This removes a "post-extraction" alteration to FPL'22 buffer
placement's results which removed a transparent slot from the placement
results if one was requested by the MILP along an opaque slot. The
rationale for this was to compensate for legacy Dynamatic's "incorrect"
placement of an elastic buffer when a single opaque slot is requested.
However, it makes more sense to respect the MILP's decisions and fix the
actual problem when integrating our new backend in the future.