[Buffers] Fix nullptr deref. with mixed-domain timing model
This fixes a crash in the mixed-domain unit path constraints code that
would dereference a nullptr timing model if the unit under consideration
had no timing model attached. Now, if the timing model for a given unit
does not exist, we simply assume that all combination delays in the
mixed domain are 0.