[NameAnalysis] Stop relying on op. pointers for name checks
This fixes a critical bug in the naming analysis which resulted in some
operations being left unnamed in the IR after naming walks at the end of
each pass and in incorrect return values from `getName` method calls.

The cause for the bug was that the analysis used to store a mapping
between `Operation*` and `std::string`'s to remember operation to name
mappings. This map was queried internally to identify the name of known
operations in amortized constant time. Importantly, mappings would
remain even after operation deletion. This could cause the analysis to
incorrectly believe that unnamed operations were named if their memory
address was identical to an old deleted operation (MLIR must reuse
previously-allocated memory space for performance reasons internally),
causing inconsistencies between the actual IR state and the analysis's
internal state.

The analysis no longer relies on operation pointer addresses for
queries, which fixes the bug. `getName` remains an amortized constant
time call (we now directly lookup the name atribute on the operation
instead of performing a map query).