[NameAnalysis] Use empty `StringRef`s instead of empty strings
This commit changes the signature of the free `getUniqueName` function
so that it returns an empty `StringRef` instead of an empty string to
indicate that the queried operation does not have a name. This avoids a
potentially unnecessary string copy when the operation does have a name,
in which case a reference to it may be returned. For consistency, an
internal naming analysis function is also modified to rely on an empty
`StringRef` instead of an optional `StringRef` to indicate a name
abscence.