[Support] Add "Handshake CFG" analysis helpers
This commit introduces a new `dynamatic::HandshakeCFG` class that is
aimed to centralize "CFG analysis" methods on Handshake functions, that
is methods that work on the implicit CFG within Handshake functions
(which originates from an explicit cf-level CFG). Internally, the class
relies on the basic block attribute optionally attached to Handshake
operations to identify properties of the implicit CFG.

This first commit adds one method to find all non-cyclic paths between
two basic blocks and one method to identify the control signal of each
basic block in a Handshake function.