[MarkMemoryInterfaces] Self-WAW does not require LSQ
This commit improves the memory interface tagging logic so that it no
longer considers WAW dependencies between an operation and itself as
warranting an LSQ for the memory access represented by the operation.
Dataflow circuit guarantee that multiple executions of the same
operation always happen in order, so this kind of "self-WAW" dependency
is ensured naturally without requiring an LSQ to order memory accesses.

This change strictly reduces the number of LSQs we place, hoora.