[HandshakeMinimizeLSQUsage] LSQ optimization pass (#56)
This commit introduces a significant optimization pass that aims to
remove access ports from LSQs when the memory dependencies that
justified their connection to the LSQ in the first place can be proven
to be honored by the dataflow circuit itself. This can result in the
size reduction or even elimination of some LSQs, improving circuit
performance and area drastically.

This optimization technique was introduced in a paper published at
FPT'19 (https://ieeexplore.ieee.org/document/8977873) and relies on the
concept of global in-order independence (GIID) between LSQ loads and
stores to guarantee the enforcement of WAR dependencies absent LSQs
(loosely, when the data result of a load is used to compute the address
or data value of a store, then a potential WAR dependence between the
two is automatically enforced by the dataflow circuit). We create the
GIID identification logic in our CFG-related helpers so that other parts
of the repository may use it in the future, should there be a need.

This commit adds the new pass to our regular compilation script as the
first IR transformation post cf-to-handshake conversion. As a result, we
get stricly less (or smaller) LSQs in all our integration tests.

Closes #56.