[Buffers] Simple buffering respects channel constraints
This commit improves our simple buffering strategy (that does not
require solving an MILP) to make it respect channel-specific buffering
constraints and take into account internal buffers in units like the
MILP-based strategies. The strategy continues to place an opaque and
transparent buffer slot on the result of any merge-like operation
(unless explictly prevented by the channel's properties or made
redundant by internal buffers in adjacent units). Additionally, the
strategy places the minimum number of slots of each type on each
requesting channel of the dataflow circuit, ensuring that all
constraints are honored no matter what.

Since this logic is similar to what our MILP-based strategies achieve,
some common code is moved from the `BufferPlacementMILP` class
definition to the `BufferingSupport` compile unit so that we can also
use it in this simpler strategy.